+++ app/(dashboard)/strategies/_ui/strategy-list/styles.module.scss
@@ -1,0 +1,3 @@
+.pagination {
+  margin-bottom: 24px;
+}
