+++ app/(dashboard)/strategies/page.module.scss
@@ -2,10 +2,6 @@
   margin-top: 80px;
 }
 
-.pagination {
-  margin-bottom: 24px;
-}
-
 .search-bar {
   width: 100%;
   background-color: $color-orange-200;
